==> 01 - Data Cleaning.ipynb <==
# Project Phase 1: Data Cleaning
In this stage, the dataset will be downloaded, previewed, and assesed for potential data formatting problems. It is in this stage where missing data, incorrect values and data forats will be dealt with.

## Importing and Previewing the dataset
We will start by importing `Pandas` and `Numpy`. We will use these package for *data frame manipulation* and *vectorized array operation*.
Once imported, we will proceed to loading the dataset in a `Pandas.DataFrame` and then preview it.

import pandas as pd
import numpy as np

# Display all rows and columns during previewing of dataframes
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

# Define the saved location of the dataset here
DATASET_URI = "./dataset/WA_Fn-UseC_-HR-Employee-Attrition.csv"

# Store the dataset in a dataframe
DATASET = pd.read_csv(DATASET_URI, index_col="EmployeeNumber")
DATASET.head()

## Categorizing Attributes
To make sure that each of the attributes have the correct data type format, we need to manually categorize each feature according to their correct type.
This *correctness* is subjective though and depends upon interpretation. The `ATTRS_CATEGORY` variable below contains the mapping of each attribute name of the dataset to their target data type.

Mapping: 
- `NUM`: Continuous/Counted Numerical Data
- `ORD`: Ordinal / Rating Data
- `NOM`: Nominal Categorical Data
- `TAR`: Target data

# Maps the column to their respective category.

# Mapping: 
#   NUM: Continuous/Counted Numerical Data
#   ORD: Ordinal / Rating Data
#   NOM: Nominal Categorical Data
#   TAR: Target data

ATTRS_CATEGORY = {
  'Age': 'NUM',
  'Attrition': 'TAR',
  'BusinessTravel': 'NOM',
  'DailyRate': 'NUM',
  'Department': 'NOM',
  'DistanceFromHome': 'NUM',
  'Education': 'ORD',
  'EducationField': 'NOM',
  'EmployeeCount': 'NUM',
  'EnvironmentSatisfaction': 'ORD',
  'Gender': 'NOM',
  'HourlyRate': 'NUM',
  'JobInvolvement': 'ORD',
  'JobLevel': 'ORD',
  'JobRole': 'NOM',
  'JobSatisfaction': 'ORD',
  'MaritalStatus': 'NOM',
  'MonthlyIncome': 'NUM',
  'MonthlyRate': 'NUM',
  'NumCompaniesWorked': 'NUM',
  'Over18': 'NOM',
  'OverTime': 'NOM',
  'PercentSalaryHike': 'NUM',
  'PerformanceRating': 'NUM',
  'RelationshipSatisfaction': 'ORD',
  'StandardHours': 'NUM',
  'StockOptionLevel': 'ORD',
  'TotalWorkingYears': 'NUM',
  'TrainingTimesLastYear': 'NUM',
  'WorkLifeBalance': 'ORD',
  'YearsAtCompany': 'NUM',
  'YearsInCurrentRole': 'NUM',
  'YearsSinceLastPromotion': 'NUM',
  'YearsWithCurrManager': 'NUM'
}

# List of attributes categorized
ATTRS_ALL = DATASET.columns
ATTRS_NUM = [attr for attr in ATTRS_ALL if ATTRS_CATEGORY[attr] == "NUM"]
ATTRS_ORD = [attr for attr in ATTRS_ALL if ATTRS_CATEGORY[attr] == "ORD"]
ATTRS_NOM = [attr for attr in ATTRS_ALL if ATTRS_CATEGORY[attr] == "NOM"]
ATTR_TAR = [attr for attr in ATTRS_ALL if ATTRS_CATEGORY[attr] == "TAR"]

## Dealing with Inconsistent Data Format
Upon checking on the dataset preview, it can be observed that the `Attrition` attribute was set as a *target* or dependent variable.
In order to conduct regression/classification tasks, this attribute needs to be encoded numerically. We can encode it using *ordinal-encoding* technique with the help of `Pandas.DataFrame.map()` method.

# Pre-processing Attrition Attribute since it is an ordinal data but contains string.
DATASET["Attrition"] = DATASET["Attrition"].map({
  "No": 0,
  "Yes": 1,
})

DATASET.head()

The next task is to make sure that each of the attribute category have the correct data type. We can achieve this using the `Pandas.DataFrame.astype()` method.

# Define the data types of each attribute category
DATASET[ATTRS_NUM] = DATASET[ATTRS_NUM].astype("float")
DATASET[ATTRS_ORD] = DATASET[ATTRS_ORD].astype("int")
DATASET[ATTRS_NOM] = DATASET[ATTRS_NOM].astype("object")
DATASET[ATTR_TAR] = DATASET[ATTR_TAR].astype("int")

## Re-arranging the Order of Features
Now, the inconsistencies of data format was dealtwith. We are now ready to proceed to the next phase.
The code below will rearrange the order of the attributes such that the numerical attributes goes first, then the ordinal ones, then the nominal ones, and finally the target variable. We do this so that we can easily separate the Features and targets later using `list`/`ndarray` indexing.

# Rearrange attributes by category
ATTRS_ALL_REARRANGED = [attr for attr_set in [ATTRS_NUM, ATTRS_ORD, ATTRS_NOM, ATTR_TAR] for attr in attr_set]

DATASET = DATASET[ATTRS_ALL_REARRANGED]

pd.DataFrame({
  "Attribute": np.array(ATTRS_ALL_REARRANGED)
})

## Exporting the Dataset
Now that the initial cleaning was conducted, we can now export our dataset into a `csv` file.

# Export the dataset
DATASET.to_csv("dataset/cleaned/Dataset.csv")

# Export the categorized attributes
pd.Series(ATTR_TAR).to_csv("dataset/constants/ATTR_TAR.csv")
pd.Series(ATTRS_NUM).to_csv("dataset/constants/ATTRS_NUM.csv")
pd.Series(ATTRS_ORD).to_csv("dataset/constants/ATTRS_ORD.csv")
pd.Series(ATTRS_NOM).to_csv("dataset/constants/ATTRS_NOM.csv")
pd.Series(ATTRS_ALL_REARRANGED).to_csv("dataset/constants/ATTRS_ALL_REARRANGED.csv")

## Closing
This concludes the data cleaning phase. In this phase, we barely do anything. Missing data was not checked on this phase, however. If you visit the dataset's kaggle website, you can find that there were no missing data in this dataset. Now our data can be readily used to plot multiple charts automatically at the next phase -- [Data Visualization](./01%20-%20Data%20Visualization.ipynb).
