==> 03 - Data Preprocessing.ipynb <==
# Project Phase 3: Data Preprocessing
In this stage, the dataset will be formatted so that it can be used in machine learning model. That is, all data should be encoded numerically. At this stage, one-hot encoding and ordinal encoding will be used to encode non-numeric data. In this notebook, we will also execute the feature selection phase.

## Loading the Cleaned Dataset
Now, let's start by loading our cleaned dataset into a `Pandas.DataFrame` object.

import pandas as pd
import numpy as np

# Import the Feature Names, convert them to numpy and them flatten them into one-dimension.
ATTRS_NUM = pd.read_csv("dataset/constants/ATTRS_NUM.csv", index_col=0).to_numpy().flatten()

DATASET = pd.read_csv("dataset/cleaned/Dataset.csv", index_col="EmployeeNumber")
DATASET.head()

## Splitting the Dataset into Features and Targets
Now, we will split the columns of our dataset into two: *Features* (or matrix of independent variable `X`), and *Target* (or the vector of dependent variable `y`). Now we can appreciate the column re-arrangement step we conducted on the [data cleaning phase](./01%20-%20Data%20Cleaning.ipynb). Using `Pandas.DataFrame.iloc[]` attribute, we can split our dataset into two slices.

# Define Feature Matrix and Target Vector to be inputted on machine learning models.
X = DATASET.iloc[:, 0:-1]
y = DATASET.iloc[:, -1]

# Conduct One-Hot Encoding on the Nominal Data of the feature matrix
X = pd.get_dummies(data=X, drop_first=True)

# Save the column names of the newly encoded dataset
ATTRS_ENCODED = X.columns

# Preview the Features
X.head()

# Preview the Target
y.head()

## Scaling the Features
If you revisit the histograms from the [data visualization phase](./02%20-%20Data%20Visualization.ipynb), you can observe that the range of our attributes were highly skewed. For instance, the `Age` attribute ranges from `20` to `60`, but the `DailyRate` ranges from `200` to `1400`. This skewed range can affect the speed and accuracy of our machine learning models. We can deal with it using `sklearn.preprocessing.StandardScaler` or other types of feature scalers.

from sklearn.preprocessing import StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.model_selection import train_test_split

# Convert the X, and y DataFrames into NDArray
X = X.values
y = y.values

# Target columns to be scaled
COLUMNS_TO_BE_SCALED = [i for i in range(0,len(ATTRS_NUM))]

# Define the column transformer with standard scaler targetted to columns defined in COLUMNS_TO_BE_SCALED.
column_transformer = ColumnTransformer(
  [("Standard Scaler", StandardScaler(),  COLUMNS_TO_BE_SCALED),],
  remainder="passthrough"
)

# Scale the whole feature set
X_scaled = column_transformer.fit_transform(X)

## Selecting the Features
Now our dataset was scaled. The range of the values of each numeric attribute of our dataset were not skewed. Let's now deal with the problem of selecting features. Similar with skewed features, using unnecessary features also slows down and affects machine learning model's performance. To deal with it, let's objectively determine the important features for our data modelling. We can do this with the help of `sklearn.feature_selection.SelectFromModel`, and using `sklearn.ensembleExtraTreesClassifier` as model for selecting features.

# Feature Selection
from sklearn.feature_selection import SelectFromModel
from sklearn.ensemble import ExtraTreesClassifier

# Fit the classifier to all of the dataset
classifier = ExtraTreesClassifier(n_estimators=50)
classifier.fit(X_scaled, y)

# Select only the important features
model = SelectFromModel(classifier, prefit=True)


X_scaled_feature_selected = model.transform(X_scaled)

# Save the names of the selected features
model.feature_names_in_ = ATTRS_ENCODED
ATTRS_SELECTED = model.get_feature_names_out()

# Preview the selected features
pd.DataFrame({
  "Selected Features": ATTRS_SELECTED
})

Now using a decision tree, out of 46 features we initially had, only 23 of them were found to be usefull in modelling. Now our models will run much faster and with better performance.

## Balancing the Dataset
Going back to the [data visualization phase](./02%20-%20Data%20Visualization.ipynb), it can be observed that the amount of data for employees that stayed at the company is way much larger than those that left. This imbalance in the target variable might result to high *accuracy score*, but poor *recall score*. To deal with this, we can balance the data by using oversampling techniques such as SMOTE, where a synthetic minority data will be generated automatically to match the number of majority data.

# Balancing the Dataset using SMOTE
from imblearn.over_sampling import SMOTE

smote = SMOTE(random_state=22)

X_scaled_resampled, y_resampled = smote.fit_resample(X_scaled_feature_selected, y)
# X_scaled_resampled, y_resampled = smote.fit_resample(X_scaled, y)

## Splitting the Dataset Into Training Set and Testing Set
To evaluation the performance of our machine learning models, we need to test it with data they have not seen so far. To do this, we will train them in a training set, then evaluation their performance on a testing set. To split the data, we can use the `sklearn.model_selection.train_test_split()` method.

# Split the features and targets into a training set and test set.
X_train, X_test, y_train, y_test = train_test_split(X_scaled_resampled, y_resampled, test_size=0.20)

## Exporting the Preprocessed Data
Finaly! We are now ready to proceed to the exciting part -- modelling. Our data is now prepared. We will conclude this phase by exporting our training, and testing set into a CSV file so that it can be readily used by other machine learning models.

# Export the training and testing set.
pd.DataFrame(X_train).to_csv("dataset/preprocessed/Features_Training_Set.csv")
pd.DataFrame(X_test).to_csv("dataset/preprocessed/Features_Testing_Set.csv")
pd.Series(y_train).to_csv("dataset/preprocessed/Target_Training_Set.csv")
pd.Series(y_test).to_csv("dataset/preprocessed/Target_Testing_Set.csv")

# Export the names of the selected features
pd.Series(ATTRS_SELECTED).to_csv("./dataset/constants/ATTRS_SELECTED.csv")
